drawing-containers/requirements.txt
boto3
ipykernel
pillow